Development

+++ database.py
@@ -1,6 +1,6 @@
 import datetime
 from contextlib import contextmanager
-from sqlalchemy import Column, ForeignKey, Integer, Text, String, create_engine, DateTime, Boolean
+from sqlalchemy import Column, ForeignKey, Integer, Text, String, create_engine, DateTime, Boolean, Float
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship, sessionmaker
 import os
@@ -33,7 +33,7 @@ class Post(Base):
     date = Column(DateTime, nullable=False)
     author_id = Column(String(32), nullable=False)
     source = Column(String(50))
-    sentiment = Column(Integer, nullable = True)
+    sentiment = Column(Float, nullable = True)
     __mapper_args__ = {
         'polymorphic_identity': 'post',
         'polymorphic_on': source
